pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "cloudconfig"
version = "0.1.2"
description = ""
authors = [
    "Adam Frank <afrank@mozilla.com>",
    "Bryan Sieber <bsieber@mozilla.com>",
    "Chris Valaas <cvalaas@mozilla.com>",
    "Jorge Spiropulo <jspiropulo@mozilla.com>"
]
packages = [
    { include = "dpm" }
]

[[tool.poetry.source]]
name = "mozilla-it"
url = "https://dp2-prod.appspot.com/pypi"

[tool.poetry.dependencies]
# These packages are mandatory and form the core of this package’s distribution.
python = ">=3.7"
pyjson = "*"
pyyaml = "*"
jsonmerge = "*"
google-cloud-firestore = "*"
cloudsecrets = { git = "https://github.com/mozilla-it/cloudsecrets.git", branch = "master" }
python3-protobuf = "^2.5.0"
protobuf = "^3.12.2"

[tool.poetry.dev-dependencies]
pre-commit = "*"
tox = "*"
behave = "*"
sphinx = "*"
pytest = "*"
pytest-cov = "*"
pytest-mock = "*"
mock = "*"
mockito = "*"

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = True
envlist = py37

[testenv]
deps =
    behave
commands =
    behave dpm/tests/bdd
"""


[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
